Step 9.13: Extract BackupMenuModal, RestoreConfirmModal, AutoBackupManagerModal
